started
have got the function asking the name of the user so I can put it on the leaderboard. then asking them to gtuess a number between 1-20. the computer uses the code of math.floor(math.random() to randomly generate a code. if the user guesses the incorrect number the code will say their incorrect and asked them again. if they are correct it will tell them they are correct and stop the code. I still need to get the count +1 to work.