invalid input testing
I tested the code to make sure that if I input a number below 1 and over 20, it should come up wityh an invalid input message. I put the number 85 into the guesser and I got the out put of that is not an valid iunput. I put 0.8 into the guesser and got the output of that is not a valid input. this is good, and what I want the code to do.

TTYL