pushed amount of tries and name
pushed amount of tries and name to the scoreboard and now it remembers the the amount of tries and name. I inputed Andra and 7 was the correct number, I guessed iu=t in 2 tries and it said that. I then played again as emma and it also push that to the score board and remembers it till the page is refreshed