boundary testing - invalid inputs
tested the function of user guess, and the invalid inputs, 1st I guess 4 which was incorrect, 2nd I guess 87 which was incorrect and invalid, 3rd I guessed 16 which was incorrect, 4th I guessed 1 which was correct. Since my 2nd gues of 87 was invalid and it told me that it was which is good, I did not want it to add +1 to the count and it didnt so that is great.

TTYL