boundary testing
I inputed the number 77 which is not between the expected input of 1-20. the cimpute them told the user that 77 is an invalid input and does not count that as a guess which is what I want.
when I input 3, the computer continues with the code becuase 3 is between 1-20. as expected