Simplied the code
is easier to read with multiple functions. and has put things in the correct function. Used a validtor and return code for the username. if it has digits. if it has didgits the code will assign 'false' to the username and later inthe code when they input their name eg:Emma88, if it has numbers it will say invaid and the code will stop because that is what 'return' does. if the username has no digits the code will chekc to make sure and will continue as per normal guessing the number.